add styling to components
