--- src/App.js
@@ -1,7 +1,0 @@
-import "./App.css";
-
-function App() {
-  return;
-}
-
-export default App;
