+++ src/components/TodosList.js
@@ -1,8 +1,6 @@
-import TodoItem from '@/components/TodoItem';
+import TodoItem from "./TodoItem";
 
-const TodosList = ({
-  todosProps, handleChange, delTodo, setUpdate,
-}) => (
+const TodosList = ({ todosProps, handleChange, delTodo, setUpdate }) => (
   <ul>
     {todosProps.map((todo) => (
       <TodoItem
