+++ src/components/TodosLogic.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { v4 as uuidv4 } from "uuid";
-import InputTodo from "@/InputTodo";
-import TodosList from "@/TodosList";
+import InputTodo from "./InputTodo";
+import TodosList from "./TodosList";
 
 const TodosLogic = () => {
   function getInitialTodos() {
